# docs/swagger.yaml
definitions:
...
  model.Menu:
    properties:
      created_at:
        type: string
      id:
        type: integer
      parent_id:
        description: 父级ID
        type: integer
      show:
        description: 是否展示 1是 2否
        type: integer
      sort:
        description: 菜单排序
        type: integer
      title:
        description: 菜单名称
        type: string
      updated_at:
        type: string
      url:
        description: 路由地址
        type: string
    type: object
...
  model.Permissions:
    properties:
      created_at:
        type: string
      http_method:
        description: 请求方法  空为任何方法 ，GET，POST，PUT，DELETE，PATCH，OPTIONS
        type: string
      http_path:
        description: 请求路径
        type: string
      id:
        type: integer
      name:
        description: 权限英文名
        type: string
      parent_id:
        description: 父级ID
        type: integer
      tag:
        description: 权限名称
        type: string
      updated_at:
        type: string
    type: object
...
  response.Code:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
info:
  contact:
    email: xiangfudada@163.com
    name: Xfdada
  title: ICQQG接囗文档
  version: "1.0"
paths:
...
  /api/web/menu:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 排序升序
        in: formData
        name: sort
        required: true
        type: integer
      - description: 菜单标题
        in: formData
        name: title
        required: true
        type: string
      - description: 菜单路由
        in: formData
        name: url
        required: true
        type: string
      - description: 是否显示 1是 2否
        in: formData
        name: show
        required: true
        type: integer
      - description: 父级ID
        in: formData
        name: parent_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 添加菜单信息
      tags:
      - 后端菜单模块
  /api/web/menu/{id}:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 删除菜单信息
      tags:
      - 后端菜单模块
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":model.Menu}'
          schema:
            $ref: '#/definitions/model.Menu'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取菜单信息
      tags:
      - 后端菜单模块
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 排序升序
        in: formData
        name: sort
        type: integer
      - description: 菜单标题
        in: formData
        name: title
        type: string
      - description: 菜单路由
        in: formData
        name: url
        type: string
      - description: 是否显示 1是 2否
        in: formData
        name: show
        type: integer
      - description: 父级ID
        in: formData
        name: parent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 修改菜单信息
      tags:
      - 后端菜单模块
  /api/web/menuList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.Menu}'
          schema:
            items:
              $ref: '#/definitions/model.Menu'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 菜单列表
      tags:
      - 后端菜单模块
  /api/web/menuTree:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]map[string]interface{}}'
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取菜单层级信息
      tags:
      - 后端菜单模块
  /api/web/permission:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 权限名
        in: formData
        name: name
        required: true
        type: string
      - description: 权限中文名
        in: formData
        name: tag
        required: true
        type: string
      - description: 请求路径
        in: formData
        name: http_path
        required: true
        type: string
      - description: 空为any 任何方法，get,put,post
        in: formData
        name: http_method
        type: string
      - description: 父级ID
        in: formData
        name: parent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 添加权限信息
      tags:
      - 后端权限模块
  /api/web/permission/{id}:
    delete:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 删除权限信息
      tags:
      - 后端权限模块
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":model.Permission}'
          schema:
            $ref: '#/definitions/model.Permissions'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取权限信息
      tags:
      - 后端权限模块
    put:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限名
        in: formData
        name: name
        type: string
      - description: 权限中文名
        in: formData
        name: tag
        type: string
      - description: 请求路径
        in: formData
        name: http_path
        type: string
      - description: 空为any 任何方法，get,put,post
        in: formData
        name: http_method
        type: string
      - description: 父级ID
        in: formData
        name: parent_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"msg":"success"}'
          schema:
            $ref: '#/definitions/response.Code'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 修改权限信息
      tags:
      - 后端权限模块
  /api/web/permissionList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]model.Permission}'
          schema:
            items:
              $ref: '#/definitions/model.Permissions'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 顶级权限列表
      tags:
      - 后端权限模块
  /api/web/permissionTree:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":[]map[string]interface{}}'
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/response.Code'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Code'
      summary: 获取权限层级信息
      tags:
      - 后端权限模块
...
swagger: "2.0"